requirements.txt
asttokens==2.0.5
backcall==0.2.0
captum==0.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
entrypoints==0.4
executing==0.8.3
fonttools==4.32.0
idna==3.3
ipykernel==6.12.1
ipython==8.2.0
jedi==0.18.1
joblib==1.1.0
jupyter-client==7.2.2
jupyter-core==4.9.2
kiwisolver==1.4.2
matplotlib==3.5.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
numpy==1.22.3
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
prompt-toolkit==3.0.29
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.11.2
pyparsing==3.0.8
python-dateutil==2.8.2
pyzmq==22.3.0
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
stack-data==0.2.0
threadpoolctl==3.1.0
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchvision==0.12.0+cu113
tornado==6.1
traitlets==5.1.1
typing_extensions==4.1.1
urllib3==1.26.9
wcwidth==0.2.5